Remove multiprocessing of plots

+++ ugali/pipeline/run_05.0_followup.py
@@ -93,7 +93,7 @@ def do_plot(args):
         logger.warning("Couldn't find %s; skipping..."%srcfile)
         return
     if not exists(samfile):
-        logger.warning("Couldn't find %s; skipping..."%samples)
+        logger.warning("Couldn't find %s; skipping..."%samfile)
         return
 
     config = ugali.utils.config.Config(config)
@@ -108,15 +108,16 @@ def do_plot(args):
     plt.savefig(outfile,bbox_inches='tight',dpi=60)
     plt.close()
 
-    data = pyfits.open(memfile)[1].data
+    if exists(memfile):
+        data = pyfits.open(memfile)[1].data
 
-    plt.figure()
-    kernel,isochrone = source.kernel,source.isochrone
-    ugali.utils.plotting.plotMembership(config,data,kernel,isochrone)
-    outfile = samfile.replace('.npy','_mem.png')
-    logger.info("  Writing %s..."%outfile)
-    plt.savefig(outfile,bbox_inches='tight',dpi=60)
-    plt.close()
+        plt.figure()
+        kernel,isochrone = source.kernel,source.isochrone
+        ugali.utils.plotting.plotMembership(config,data,kernel,isochrone)
+        outfile = samfile.replace('.npy','_mem.png')
+        logger.info("  Writing %s..."%outfile)
+        plt.savefig(outfile,bbox_inches='tight',dpi=60)
+        plt.close()
 
     
 def run(self):
@@ -170,8 +171,9 @@ def run(self):
 
     if 'plot' in self.opts.run:
         logger.info("Running 'plot'...")
-        pool = Pool(maxtasksperchild=1)
-        pool.map(do_plot,args)
+        #pool = Pool(maxtasksperchild=1)
+        #pool.map(do_plot,args)
+        map(do_plot,args)
 
     if 'collect' in self.opts.run:
         logger.info("Running 'collect'...")
